requirements.txt
cycler==0.10.0
imagecodecs==2019.12.31
imagecodecs-lite==2019.12.3
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.1
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.14.0
tifffile==2019.7.26.2
tiffile==2018.10.18
